separate compositions
