+++ src/components/Main/MainView/MainViewHeader/use/channelCreateModal.ts
@@ -1,0 +1,14 @@
+import store from '@/store'
+import { ChannelId } from '@/types/entity-ids'
+
+const useChannelCreateModal = (props: { channelId: ChannelId }) => {
+  const openChannelCreateModal = () => {
+    store.dispatch.ui.modal.pushModal({
+      type: 'channel-create',
+      parentChannelId: props.channelId
+    })
+  }
+  return { openChannelCreateModal }
+}
+
+export default useChannelCreateModal
