+++ src/components/Main/MainView/MainViewHeader/use/notificationModal.ts
@@ -1,0 +1,14 @@
+import store from '@/store'
+import { ChannelId } from '@/types/entity-ids'
+
+const useNotificationModal = (props: { channelId: ChannelId }) => {
+  const openNotificationModal = () => {
+    store.dispatch.ui.modal.pushModal({
+      type: 'notification',
+      channelId: props.channelId
+    })
+  }
+  return { openNotificationModal }
+}
+
+export default useNotificationModal
